Reform 'all_day_event' private interface:
now we consider all day events only those where the time component of both start and end date is set at midnight. This is more restrictive that the previous criteria, which would allow also events where start and end date were equals, or the date component was different, regardless of time, to be considered all day events.
In practice, the user interface would already make so that an event created as "all day event" would be set with midnight as the time for start and end date.

calendar::item::get would already return time_p in the array to indicate whether this was an event with time or not. No time would mean "all day event". For convenience, we now return also the straight all_day_event_p flag in the array.